pyproject.toml
[project]
name = "ozonpricetracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Fedor Smirnov",email = "f.smirnov@innopolis.university"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
package-mode = false


dependencies = [
    "mutmut==2.4.4",
    "bandit (>=1.8.3,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
    "flake8-pyproject (>=1.2.3,<2.0.0)",
    "ruff (>=0.11.8,<0.12.0)",
    "coverage (>=7.8.0,<8.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
]

[tool.poetry.group.frontend.dependencies]
streamlit = "==1.45.0"

[tool.poetry.group.backend.dependencies]
seleniumbase = "4.29.0"
undetected-chromedriver = "3.5.0"

[tool.poetry]
packages = [
    { include = "app", from = "." },
    { include = "test", from = "."}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.dotenv]
ignore = "false"
location = "secrets.env"